+++ lib/runners/processor.rb
@@ -218,15 +218,10 @@ module Runners
     class InvalidXML < SystemError; end
 
     def read_xml(text)
-      root = REXML::Document.new(text).root
-      if root
-        root
-      else
-        message = "Invalid XML: #{text.inspect}"
-        trace_writer.error message
-        raise InvalidXML, message
+      Nokogiri::XML(text) do |config|
+        config.strict
       end
-    rescue REXML::ParseException => exn
+    rescue Nokogiri::XML::SyntaxError => exn
       message =
         if exn.cause.instance_of? RuntimeError
           exn.cause.message