+++ lib/runners/processor/checkstyle.rb
@@ -57,7 +57,7 @@ module Runners
       target = Array(config_linter[:target] || config_linter[:dir] || DEFAULT_TARGET)
       capture3(analyzer_bin, *checkstyle_args, *target)
 
-      xml_root =
+      xml_doc =
         begin
           read_report_xml
         rescue InvalidXML
@@ -66,7 +66,7 @@ module Runners
         end
 
       Results::Success.new(guid: guid, analyzer: analyzer).tap do |result|
-        construct_result(xml_root) do |issue|
+        construct_result(xml_doc) do |issue|
           result.add_issue(issue)
         end
       end
@@ -98,14 +98,14 @@ module Runners
       end
     end
 
-    def construct_result(xml_root)
+    def construct_result(xml_doc)
       ignored_severities = Array(config_linter[:ignore])
 
-      xml_root.each_element("file") do |file|
+      xml_doc.root.elements.each do |file|
         file_name = file[:name] or raise "Invalid file: #{file.inspect}"
         path = relative_path file_name
 
-        file.each_element do |error|
+        file.elements.each do |error|
           case error.name
           when "error"
             severity = error[:severity]
@@ -125,7 +125,7 @@ module Runners
               schema: SCHEMA.issue,
             )
           when "exception"
-            exception = error.text or raise "Required exception: #{error.inspect}"
+            exception = error.content or raise "Required exception: #{error.inspect}"
             add_warning exception, file: path.to_s
           end
         end