+++ lib/runners/processor/cppcheck.rb
@@ -152,7 +152,7 @@ module Runners
         return Results::Failure.new(guid: guid, analyzer: analyzer)
       end
 
-      xml_root =
+      xml_doc =
         begin
           read_report_xml
         rescue InvalidXML
@@ -160,19 +160,19 @@ module Runners
         end
 
       Results::Success.new(guid: guid, analyzer: analyzer).tap do |result|
-        parse_result(xml_root) do |issue|
+        parse_result(xml_doc) do |issue|
           result.add_issue issue
         end
       end
     end
 
     # @see https://github.com/danmar/cppcheck/blob/master/man/manual.md#xml-output
-    def parse_result(xml_root)
-      xml_root.each_element("errors/error") do |err|
+    def parse_result(xml_doc)
+      xml_doc.search("errors/error").each do |err|
         id = err[:id] or raise "Required id: #{err.inspect}"
         msg = err[:msg] or raise "Required msg: #{err.inspect}"
 
-        err.each_element("location") do |loc|
+        err.search("location").each do |loc|
           file = loc[:file] or raise "Required file: #{loc.inspect}"
           yield Issue.new(
             id: id,