+++ lib/runners/processor/pmd_java.rb
@@ -100,14 +100,14 @@ module Runners
     def construct_result
       # https://github.com/pmd/pmd.github.io/blob/8b0c31ff8e18215ed213b7df400af27b9137ee67/report_2_0_0.xsd
 
-      read_report_xml.each_element do |element|
+      read_report_xml.root.elements.each do |element|
         case element.name
         when "file"
           filename = element[:name] or raise "Unexpected element: #{element.inspect}"
           path = relative_path(filename)
 
-          element.each_element("violation") do |violation|
-            message = violation.text or raise "required message: #{violation.inspect}"
+          element.elements.each do |violation|
+            message = violation.content or raise "required message: #{violation.inspect}"
 
             yield Issue.new(
               path: path,